% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{calculate_u_statistics_six}
\alias{calculate_u_statistics_six}
\title{Compute U-statistics HOIF-type from Order 2 to 6 in pure R code.}
\usage{
calculate_u_statistics_six(Vector_1, Vector_2, A1, A2, A3, A4, A5)
}
\arguments{
\item{Vector_1}{Numeric column vector of length \eqn{n}.}

\item{Vector_2}{Numeric column vector of length \eqn{n}.}

\item{A1, A2, A3, A4, A5}{Numeric \eqn{n \times n} kernel matrices of the same dimension.}
}
\value{
A named list containing numeric U-statistic estimates:
\describe{
  \item{U2}{Second-order U-statistic}
  \item{U3}{Third-order U-statistic}
  \item{U4}{Fourth-order U-statistic}
  \item{U5}{Fifth-order U-statistic}
  \item{U6}{Sixth-order U-statistic}
}
}
\description{
This function serves as a core computational component in higher-order
influence function (HOIF) estimators in pure R code.
}
\details{
Internally, the function constructs kernel matrices for orders 2 through 6
using recursive matrix operations and removes diagonal contributions to
ensure degenerate U-statistics.


All diagonal elements of intermediate kernel matrices are removed to avoid
self-interactions. Matrix multiplications are performed via
`eigenMapMatMult()` and element-wise products via `hadamard()`.
The exact formula of the output is:
\deqn{
\mathbb{U}_{n,m} = \frac{1}{\binom{n}{m} m!}
\sum_{i_1 \ne \cdots \ne i_m} Vector_1[i_1] \cdot A1[i_1,i_2] \cdot A1[i_2,i_3] \cdots A1[i_{m-1},i_{m}] \cdot Vector_2[i_{m}]
}
}
