% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_hoif_setup}
\alias{check_hoif_setup}
\title{Check HOIF environment status}
\usage{
check_hoif_setup()
}
\value{
Prints status message and invisibly returns TRUE/FALSE
}
\description{
Checks whether the Python environment is properly configured for HOIF.
}
\examples{
\dontrun{
check_hoif_setup()
}

}
