% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_python_env}
\alias{check_python_env}
\title{Environment Setup Functions for HOIF Package}
\usage{
check_python_env()
}
\value{
Logical indicating availability
}
\description{
Functions for setting up and checking the Python environment required
for computing U-statistics.
}
\author{
Xingyu Chen
}
\keyword{internal}
