% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{compute_basis_matrix}
\alias{compute_basis_matrix}
\title{Compute basis projection matrices}
\usage{
compute_basis_matrix(Z, A, Omega1, Omega0)
}
\arguments{
\item{Z}{Basis matrix (n x k)}

\item{A}{Treatment vector (n x 1)}

\item{Omega1}{Inverse Gram matrix for treatment group}

\item{Omega0}{Inverse Gram matrix for control group}
}
\value{
List with B1 and B0 (projection matrices)
}
\description{
Compute basis projection matrices
}
\examples{
n <- 100
Z <- cbind(1, matrix(rnorm(n * 2), n, 2))
A <- rbinom(n, 1, 0.5)
Omega <- compute_gram_inverse(Z, A)
B <- compute_basis_matrix(Z, A, Omega$Omega1, Omega$Omega0)
dim(B$B1)

}
