% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{compute_residuals}
\alias{compute_residuals}
\title{Compute residuals for both treatment groups}
\usage{
compute_residuals(A, Y, mu1, mu0, pi)
}
\arguments{
\item{A}{Treatment vector (0 or 1)}

\item{Y}{Outcome vector}

\item{mu1}{Predicted outcomes under treatment (mu(1, X))}

\item{mu0}{Predicted outcomes under control (mu(0, X))}

\item{pi}{Propensity scores}
}
\value{
List with R1, r1, R0, r0
}
\description{
Compute residuals for both treatment groups
}
