% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{diag_col_sum}
\alias{diag_col_sum}
\title{Diagonal Matrix of Column Sums}
\usage{
diag_col_sum(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A diagonal matrix.
}
\description{
Creates a diagonal matrix whose diagonal elements are the column sums
of the input matrix.
}
\keyword{internal}
