% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{.onLoad}
\alias{.onLoad}
\title{Package startup function}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
Package startup function
}
\keyword{internal}
