% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoif.R
\name{expr_list_to_einstein}
\alias{expr_list_to_einstein}
\title{HOIF Estimators for Average Treatment Effect}
\usage{
expr_list_to_einstein(expr_list)

expr_list_to_einstein(expr_list)
}
\arguments{
\item{expr_list}{A list of integer vectors, each of length 2}
}
\value{
Character string in Einstein notation

Character string in Einstein notation
}
\description{
This file implements the Higher Order Influence Function (HOIF) estimators
for Average Treatment Effect (ATE) estimation with nuisance functions.

Converts a nested list expression like [[1,2],[2,3],[3,4]] to Einstein
notation like "ab,bc,cd->".
}
\examples{
\dontrun{
expr_list_to_einstein(list(c(1,2), c(2,3), c(3,4)))  # Returns "ab,bc,cd->"
}
\dontrun{
expr_list_to_einstein([[1,2],[2,3],[3,4]])  # Returns "ab,bc,cd->"
}
}
\author{
Xingyu Chen
}
\keyword{internal}
