% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{hadamard}
\alias{hadamard}
\title{Hadamard (Element-wise) Matrix Product}
\usage{
hadamard(mat1, mat2)
}
\arguments{
\item{mat1, mat2}{Numeric matrices of identical dimensions.}
}
\value{
A matrix containing the element-wise product.
}
\description{
Computes the element-wise (Hadamard) product of two matrices of the same size.
}
\keyword{internal}
