% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{data}
\name{hoif_env}
\alias{hoif_env}
\title{Package-level Documentation and Startup}
\format{
An object of class \code{environment} of length 2.
}
\usage{
hoif_env
}
\description{
Package-level Documentation and Startup
}
\author{
Xingyu Chen
}
\keyword{datasets}
