% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{plot.hoif_ate}
\alias{plot.hoif_ate}
\title{Plot convergence of ATE estimates}
\usage{
\method{plot}{hoif_ate}(x, ...)
}
\arguments{
\item{x}{Object of class hoif_ate}

\item{...}{Additional arguments passed to plot}
}
\value{
No return value, called for its side effect of drawing a
  convergence plot of the ATE estimates against the order.
}
\description{
Plot convergence of ATE estimates
}
