% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{print.hoif_ate}
\alias{print.hoif_ate}
\title{Print method for hoif_ate objects}
\usage{
\method{print}{hoif_ate}(x, ...)
}
\arguments{
\item{x}{Object of class hoif_ate}

\item{...}{Additional arguments (unused)}
}
\value{
The input object \code{x}, invisibly. Called for its side
  effect of printing a summary of the estimates to the console.
}
\description{
Print method for hoif_ate objects
}
