% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{compute_basis_matrix}
\alias{compute_basis_matrix}
\title{Compute basis projection matrices}
\usage{
compute_basis_matrix(Z, Omega1, Omega0)
}
\arguments{
\item{Z}{Basis matrix (n x k)}

\item{Omega1}{Inverse Gram matrix for treatment group}

\item{Omega0}{Inverse Gram matrix for control group}
}
\value{
List with B1 and B0 (projection matrices)
}
\description{
Compute basis projection matrices
}
