% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{compute_gram_inverse}
\alias{compute_gram_inverse}
\title{Compute inverse of weighted Gram matrix}
\usage{
compute_gram_inverse(Z, A, method = "direct")
}
\arguments{
\item{Z}{Basis matrix (n x k)}

\item{A}{Treatment vector}

\item{method}{Character: "direct", "nlshrink", or "corpcor"}
}
\value{
List with Omega1 and Omega0 (inverse Gram matrices)
}
\description{
Compute inverse of weighted Gram matrix
}
