% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{expr_list_to_einstein}
\alias{expr_list_to_einstein}
\title{Convert nested list expression to Einstein notation}
\usage{
expr_list_to_einstein(expr_list)
}
\arguments{
\item{expr_list}{A list of integer vectors, each of length 2}
}
\value{
Character string in Einstein notation
}
\description{
Converts a nested list expression like [[1,2],[2,3],[3,4]] to Einstein
notation like "ab,bc,cd->".
}
\examples{
\dontrun{
expr_list_to_einstein([[1,2],[2,3],[3,4]])  # Returns "ab,bc,cd->"
}
}
\keyword{internal}
