% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOIF.R
\name{print.hoif_ate}
\alias{print.hoif_ate}
\title{Print method for hoif_ate objects}
\usage{
\method{print}{hoif_ate}(x, ...)
}
\arguments{
\item{x}{Object of class hoif_ate}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for hoif_ate objects
}
